Restructure security concerns
