--- src/CleanArchitecture.Infrastructure/Common/Security/CurrentUserProvider/ICurrentUserProvider.cs
@@ -1,6 +1,0 @@
-namespace CleanArchitecture.Infrastructure.Common.Security.CurrentUserProvider;
-
-public interface ICurrentUserProvider
-{
-    CurrentUser GetCurrentUser();
-}
