--- src/CleanArchitecture.Infrastructure/Configuration/JwtBearerOptionsSetup.cs
@@ -1,32 +1,0 @@
-﻿using System.Text;
-
-using CleanArchitecture.Infrastructure.Security.TokenGenerator;
-
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.Extensions.Options;
-using Microsoft.IdentityModel.Tokens;
-
-namespace CleanArchitecture.Infrastructure.Configuration;
-
-public sealed class JwtBearerOptionsSetup(IOptions<JwtSettings> jwtSettings)
-    : IConfigureNamedOptions<JwtBearerOptions>
-{
-    private readonly JwtSettings _jwtSettings = jwtSettings.Value;
-
-    public void Configure(string? name, JwtBearerOptions options) => Configure(options);
-
-    public void Configure(JwtBearerOptions options)
-    {
-        options.TokenValidationParameters = new TokenValidationParameters
-        {
-            ValidateIssuer = true,
-            ValidateAudience = true,
-            ValidateLifetime = true,
-            ValidateIssuerSigningKey = true,
-            ValidIssuer = _jwtSettings.Issuer,
-            ValidAudience = _jwtSettings.Audience,
-            IssuerSigningKey = new SymmetricSecurityKey(
-                Encoding.UTF8.GetBytes(_jwtSettings.Secret)),
-        };
-    }
-}
