+++ src/CleanArchitecture.Infrastructure/Security/CurrentUserProvider/ICurrentUserProvider.cs
@@ -1,0 +1,6 @@
+namespace CleanArchitecture.Infrastructure.Security.CurrentUserProvider;
+
+public interface ICurrentUserProvider
+{
+    CurrentUser GetCurrentUser();
+}
