+++ src/CleanArchitecture.Infrastructure/Security/PolicyEnforcer/PolicyEnforcer.cs
@@ -1,7 +1,7 @@
 using CleanArchitecture.Application.Common.Security.Policies;
 using CleanArchitecture.Application.Common.Security.Request;
 using CleanArchitecture.Application.Common.Security.Roles;
-using CleanArchitecture.Infrastructure.Common.Security.CurrentUserProvider;
+using CleanArchitecture.Infrastructure.Security.CurrentUserProvider;
 
 using ErrorOr;
